+++ Assets/Script/Structs/GameResultData.cs
@@ -1,0 +1,16 @@
+using System;
+
+[Serializable]
+public struct GameResultData
+{
+    public int Day;
+    public int MaxCombo;
+    public int Score;
+     
+    public GameResultData(int day, int maxCombo, int score)
+    {
+        Day = day;
+        MaxCombo = maxCombo;
+        Score = score;
+    }
+}
