+++ Assets/Script/Structs/GameResultData.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: f9338ae084d34efea1f2bd9903ace722
+timeCreated: 1755139555
