--- Assets/Script/UI/PopupUI/ResultItem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 24636c280993e9d479859a57f4bb0e4c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
